Reorganise docker-compose files and create peon.sh
The docker-compose files were a little mess in the root directory so
they got moved into the dockerfiles folder since from now on we use a
custom script to run docker-compose anyway.

peon.sh is that script.
This script takes into account if you want to run dev or production and
has all the main commands of docker-compose in it.
And if someone needs to run a different docker-compose command, the
custom command in peon.sh should do the trick.